Update NcFile and NcVar attribute fields when adding attributes
Calling ncputatt to add an attribute or adding global attributes when
creating a new file with NetCDF.create would write the attributes to
the file on disk, but it would not update the atts/gatts field of the
corresponding NcVar or NcFile struct. This commit changes this be-
haviour and the appropriate NcVar/NcFile field is now updated to
reflect the changes to the file on disk.