use the new NetCDF_jll; require julia 1.3 (#123)
* use the new NetCDF_jll; require julia 1.3

I wanted to give the new NetCDF_jll from JuliaPackaging/Yggdrasil#1090 a test run.

Since this relies on Artifacts being available it needs julia 1.3 minimum and therefore a minor version bump, to 0.11.

https://github.com/JuliaBinaryWrappers/NetCDF_jll.jl

Also see https://github.com/Alexander-Barth/NCDatasets.jl/pull/88

* update CI versions

* update docs projects

This adds NetCDF itself as a dependency, just like is done in PkgTemplates. That means Conda is still a dependency for now until we tag a new release.

* add NetCDF_jll to deps

* update Documenter to 0.25

With prettyurls set to false for local builds, such that you don't need to run a local server to click around.